requirements.txt
# Core dependencies
requests>=2.31.0
twilio>=8.10.0
python-dotenv>=1.0.0
PyYAML>=6.0.1

# CLI
click>=8.1.7
rich>=13.7.0

# Scheduling
schedule>=1.2.1

# Data handling
pandas>=2.1.4
pytz>=2023.3

# Email support
secure-smtplib>=0.1.1

# Logging
colorlog>=6.8.0

# Database (for storing price history)
sqlalchemy>=2.0.23

# Testing (optional)
pytest>=7.4.3
pytest-cov>=4.1.0
